service_weather/requirements.txt
Flask
Flask-RESTful
redis
python-decouple